Syslog, blocking socket, joinable threads.
* Logging
All logging was rewritten with syslog.
Temporary removed debug logging on storing to DB.

* Blocking socket
Netlink socket in daemon is now blocking.
This was done because receiving on non-blocking in infinite loop occupies whole
processor core.

* Joinable threads
Threads is now created as joinable and thread_init is waiting for it's
termination with pthread_join. This was done because main thread became zombie.
When threads are detached we can't wait on it but main thread can't terminate
correctly on pthread_exit because there are running child threads.

Maybe I'm too lame with pthreads and there are better solution for that with
detached threads.